refactor: move cli commands out
