+++ cognite_toolkit/_cdf_tk/commands/modules.py
@@ -1,8 +1,6 @@
 from __future__ import annotations
 
 import shutil
-import subprocess
-from contextlib import suppress
 from importlib import resources
 from pathlib import Path
 from typing import Optional
@@ -29,6 +27,7 @@ from cognite_toolkit._cdf_tk.commands._changes import (
     ManualChange,
     UpdateModuleVersion,
 )
+from cognite_toolkit._cdf_tk.commands._cli_commands import CLICommands
 from cognite_toolkit._cdf_tk.constants import (
     BUILTIN_MODULES,
     BUILTIN_MODULES_PATH,
@@ -428,31 +427,3 @@ class ModulesCommand(ToolkitCommand):
             )
 
         print(table)
-
-
-class CLICommands:
-    @classmethod
-    def use_poetry(cls) -> bool:
-        with suppress(Exception):
-            return shutil.which("poetry") is not None
-        return False
-
-    @classmethod
-    def use_git(cls) -> bool:
-        with suppress(Exception):
-            return shutil.which("git") is not None
-        return False
-
-    @classmethod
-    def has_initiated_repo(cls) -> bool:
-        with suppress(Exception):
-            result = subprocess.run("git rev-parse --is-inside-work-tree".split(), stdout=subprocess.PIPE)
-            return result.returncode == 0
-        return False
-
-    @classmethod
-    def has_uncommitted_changes(cls) -> bool:
-        with suppress(Exception):
-            result = subprocess.run("git diff --quiet".split(), stdout=subprocess.PIPE)
-            return result.returncode != 0
-        return False
