+++ module_upgrade/run_check.py
@@ -18,7 +18,7 @@ from rich.panel import Panel
 
 from cognite_toolkit._cdf_tk.commands import BuildCommand, DeployCommand, ModulesCommand
 from cognite_toolkit._cdf_tk.commands._changes import ManualChange
-from cognite_toolkit._cdf_tk.commands.modules import CLICommands
+from cognite_toolkit._cdf_tk.commands._cli_commands import CLICommands
 from cognite_toolkit._cdf_tk.constants import ROOT_MODULES, SUPPORT_MODULE_UPGRADE_FROM_VERSION
 from cognite_toolkit._cdf_tk.loaders import LOADER_BY_FOLDER_NAME
 from cognite_toolkit._cdf_tk.utils import CDFToolConfig, module_from_path
